backend/requirements.txt
# VN Fashion Shop - Backend Dependencies
# =======================================

# Framework
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.30.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9

# Cache & Task Queue
redis>=5.0.0
celery>=5.3.0

# Payment
stripe>=8.0.0

# Image Processing
pillow>=10.0.0

# AWS S3
boto3>=1.34.0

# HTTP Client
httpx>=0.27.0

# Rate Limiting
slowapi>=0.1.9

# Environment
python-dotenv>=1.0.0

# Email
emails>=0.6

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
factory-boy>=3.3.0
